fix(agents): follow SpatialMemory move to perception/experimental
#3275 moved dimos/perception/spatial_perception.py under
perception/experimental/; update the two lazy imports in the eval
ingest and store fixture to the new path.
